// apps/api/src/generated/prisma/models/User.ts
export type UserModel = runtime.Types.Result.DefaultSelection<Prisma.$UserPayload>
// ...
export type $UserPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "User"
  objects: {
    reservations: Prisma.$ReservationPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    email: string
    password: string
    role: $Enums.Role
    createdAt: Date
    hashedRefreshToken: string | null
  }, ExtArgs["result"]["user"]>
  composites: {}
}